Relocate verifier diagnostic/repair message builders to verifier_orchestration
Phase 2 batch 5 for Issue #682. Moves the 3 big LLM-prompt builders
(~263 LOC of bodies) out of turn.rs into verifier_orchestration.rs.

Moved (turn.rs -> verifier_orchestration.rs):

  - verifier_diagnostic_messages (123 LOC) — builds the system +
    user ConversationMessage pair sent to the diagnostic classifier
    LLM. Bundles FailurePacket / AuthorityEvidence / framework
    findings / file excerpts / SemanticFailureReport schema +
    SecuritySigil note about behavior_contract being untrusted data.

  - verifier_repair_pass_messages (113 LOC) — builds the system +
    user ConversationMessage pair sent to the repair-edit LLM.
    Bundles selected_target + target excerpt + related excerpts +
    diagnostic assessment + SemanticRepairPlan + repair_action.

  - safe_verifier_repair_file_excerpt (27 LOC) — symlink-containment-
    safe file reader producing mask_secrets-pipeline-sanitized
    bounded excerpts; helper consumed only by
    verifier_repair_pass_messages.

All 3 promoted to pub(super). RecoveryTargetHint is promoted out of
the #[cfg(test)] guard since verifier_repair_pass_messages is a
production fn.

Adjustments:

  * verifier_orchestration.rs: added imports for
    super::failure_packet::FailurePacket,
    super::repair_authority::AuthorityEvidence,
    super::repair_driver::{VERIFIER_REPAIR_PASS_MAX_FILE_BYTES,
    VERIFIER_REPAIR_PASS_MAX_FILE_EXCERPT_BYTES},
    super::repair_framework_findings::findings_for_diagnostic as
    verifier_framework_findings_for_diagnostic,
    super::repair_job::{mask_code_excerpt_preserving_patch_anchors,
    mask_secrets_headers_and_neutralize},
    super::repair_patch_validation::is_repair_path_input_safe,
    super::required_behavior::behavior_contract_payload_value,
    super::verifier_failure_signature::compact_verifier_failure_text,
    super::verifier_repair_shadow::
    verifier_repair_action_payload_for_context,
    super::verifier_repair_targeting::
    verifier_diagnostic_missing_setup_candidates.

  * turn.rs:
    - Extended use super::verifier_orchestration::{...} with the 3
      production-callable relocated fns and dropped 4 now-unused
      sibling imports (VERIFIER_REPAIR_PASS_MAX_FILE_EXCERPT_BYTES,
      verifier_repair_action_payload_for_context,
      verifier_diagnostic_missing_setup_candidates,
      behavior_contract_payload_value), plus
      verifier_file_excerpt_for_line + verifier_repair_context_line_for_path
      from production import (still test-only).
    - Updated mod progress_tests to pull
      {verifier_diagnostic_messages, verifier_file_excerpt_for_line,
      verifier_repair_pass_messages} from
      super::super::verifier_orchestration:: alongside the existing
      target_discovery_note import.
    - Fixed an include_str!-based structural test
      (`v0421_repair_runner_contract_tests::
      repair_patch_provider_uses_json_mode_and_no_think`) to read
      verifier_repair_pass_messages from verifier_orchestration.rs
      with the new `pub(super) fn ` marker.

Metrics: turn.rs 34,353 -> 34,086 (-267); verifier_orchestration.rs
697 -> 973 (+276). Cumulative Issue #682 progress:
turn.rs 34,958 -> 34,086 (-872, ~30% of Phase 2 target). Acceptance
gap: 34,086 -> 32,000 = ~2,086 LOC remaining.
